Fixed UTF-8 TextFormat output to protect against invalid UTF-8 in string fields.
This will guarantee that the output of TextFormat is always valid UTF-8.  Prior to this change, any invalid UTF-8 would escape into the TextFormat output if users enabled `SetUseUtf8StringEscaping(true)` or called `Utf8DebugString()`.

This change currently only affects users who explicitly set `SetUseUtf8StringEscaping(true)` or use `Utf8DebugString()`, but the hope is to flip the default so that this mode is enabled unless overridden.

PiperOrigin-RevId: 570750983